refactor: Move core.py to discovery/manager.py, wire is_newer
- Rename napt/core.py -> napt/discovery/manager.py to match the
  build/manager.py and upload/manager.py conventions
- Use is_newer() for the cache version check instead of == equality;
  handles v-prefix normalization and None on first run correctly
- Upgrade "Version discovered" log from verbose to info so it appears
  on first run (before any cache branch is entered)
- Remove redundant explicit strategy imports inside discover_recipe;
  napt/discovery/__init__.py loads them at import time
- Update all import paths across napt/, tests/, and docstring examples
