+++ napt/discovery/api_github.py
@@ -125,7 +125,7 @@ Example:
     From Python (using core orchestration):
         ```python
         from pathlib import Path
-        from napt.core import discover_recipe
+        from napt.discovery import discover_recipe
 
         # Automatically uses version-first optimization
         result = discover_recipe(Path("recipe.yaml"), Path("./downloads"))
@@ -149,8 +149,8 @@ from typing import Any
 
 import requests
 
-from napt.exceptions import ConfigError, NetworkError
 from napt.discovery.base import RemoteVersion
+from napt.exceptions import ConfigError, NetworkError
 
 from .base import register_strategy
