+++ napt/discovery/web_scrape.py
@@ -145,7 +145,7 @@ Example:
     From Python (using core orchestration):
         ```python
         from pathlib import Path
-        from napt.core import discover_recipe
+        from napt.discovery import discover_recipe
 
         # Automatically uses version-first optimization
         result = discover_recipe(Path("recipe.yaml"), Path("./downloads"))
@@ -172,8 +172,8 @@ from urllib.parse import urljoin
 from bs4 import BeautifulSoup
 import requests
 
-from napt.exceptions import ConfigError, NetworkError
 from napt.discovery.base import RemoteVersion
+from napt.exceptions import ConfigError, NetworkError
 
 from .base import register_strategy
