+++ tests/test_core.py
@@ -13,7 +13,7 @@ from unittest.mock import patch
 
 import pytest
 
-from napt.core import discover_recipe
+from napt.discovery.manager import discover_recipe
 from napt.exceptions import ConfigError
 
 
@@ -35,7 +35,7 @@ class TestDiscoverRecipe:
         recipe_path = create_yaml_file("recipe.yaml", recipe_data)
 
         # Mock the discovery strategy (url_download - file-first)
-        with patch("napt.core.get_strategy") as mock_get_strategy:
+        with patch("napt.discovery.manager.get_strategy") as mock_get_strategy:
             mock_strategy = mock_get_strategy.return_value
             # Ensure it doesn't have get_version_info (file-first strategy)
             del mock_strategy.get_version_info
@@ -156,11 +156,11 @@ class TestVersionFirstFastPath:
         with requests_mock.Mocker() as m:
             m.get("https://example.com/download.html", text=html_content)
 
-            with patch("napt.core.load_state") as mock_load_state:
+            with patch("napt.discovery.manager.load_state") as mock_load_state:
                 mock_load_state.return_value = state
 
-                with patch("napt.core.save_state"):
-                    with patch("napt.core.download_file") as mock_download:
+                with patch("napt.discovery.manager.save_state"):
+                    with patch("napt.discovery.manager.download_file") as mock_download:
                         result = discover_recipe(
                             recipe_path, tmp_test_dir, state_file=Path("state.json")
                         )
@@ -214,11 +214,11 @@ class TestVersionFirstFastPath:
         with requests_mock.Mocker() as m:
             m.get("https://example.com/download.html", text=html_content)
 
-            with patch("napt.core.load_state") as mock_load_state:
+            with patch("napt.discovery.manager.load_state") as mock_load_state:
                 mock_load_state.return_value = state
 
-                with patch("napt.core.save_state"):
-                    with patch("napt.core.download_file") as mock_download:
+                with patch("napt.discovery.manager.save_state"):
+                    with patch("napt.discovery.manager.download_file") as mock_download:
                         from napt.results import DownloadResult
 
                         mock_download.return_value = DownloadResult(
@@ -287,10 +287,10 @@ class TestVersionFirstFastPath:
                 headers={"Content-Length": str(len(fake_content))},
             )
 
-            with patch("napt.core.load_state") as mock_load_state:
+            with patch("napt.discovery.manager.load_state") as mock_load_state:
                 mock_load_state.return_value = state
 
-                with patch("napt.core.save_state"):
+                with patch("napt.discovery.manager.save_state"):
                     result = discover_recipe(
                         recipe_path, tmp_test_dir, state_file=Path("state.json")
                     )
